Expose token without query parameter structure at API level
The query parameter templating is good for use in templates but when accessing directly in Python we're not typically dealing with strings, we're dealing with QueryDicts or otherwise want the "raw" token more of the time.

This also means that our API of getting a token vs validating a token is more symmetric.